Use int for timeouts

+++ src/client.h
@@ -44,11 +44,11 @@ int client_set_qtype(char *qtype);
 char *client_get_qtype();
 void client_set_downenc(char *encoding);
 void client_set_compression(int up, int down);
-void client_set_dnstimeout(double, double, double, int);
+void client_set_dnstimeout(int, int, int, int);
 void client_set_lazymode(int lazy_mode);
 void client_set_windowsize(size_t, size_t);
 void client_set_hostname_maxlen(size_t i);
-void client_set_interval(double, double);
+void client_set_interval(int, int);
 
 int client_handshake(int dns_fd, int raw_mode, int autodetect_frag_size, int fragsize);
 int client_tunnel(int tun_fd, int dns_fd);
